requirements.txt
matplotlib==3.5.1
numpy==1.21.2
pandas==1.4.1
Pillow==9.2.0
seaborn==0.11.2
streamlit==1.8.1
folium==0.12.1.post1
ee==0.1.316
geemap==0.14.2
streamlit-folium==0.6.13
